Adds methods to send out each packet
Event handlers have now been created in the server, network events are actually sent when an event is triggered in the RoundManager

+++ ProjectFiles/Assets/Scripts/Game/World.cs
@@ -105,12 +105,12 @@ namespace Game
             return cars.ContainsKey(playerID);
         }
 
-        public int GetNumPlayers()
+        public ushort GetNumPlayers()
         {
-            return cars.Count;
+            return (ushort) cars.Count;
         }
 
-        public List<int> GetPlayersNotInSpace()
+        public List<ushort> GetPlayersNotInSpace()
         {
             // TODO: Implement me ;)
             throw new System.NotImplementedException();
